Move image encoding functions to Encode trait
